Rename loadout/item-utils to loadout-item-utils
