+++ src/app/loadout-builder/useEquippedHashes.ts
@@ -1,5 +1,5 @@
 import { LoadoutParameters } from '@destinyitemmanager/dim-api-types';
-import { getSubclassPlugHashes } from 'app/loadout/item-utils';
+import { getSubclassPlugHashes } from 'app/loadout/loadout-item-utils';
 import { ResolvedLoadoutItem } from 'app/loadout/loadout-types';
 import { useMemo } from 'react';
