+++ src/app/loadout/loadout-item-utils.ts
